Split osm for clarity
